requirements.txt
# CitizenGov – Version-Locked Dependencies
# Python 3.10+

streamlit==1.31.1
plotly==5.18.0
pandas==2.2.0
requests==2.31.0
pydantic==2.6.1
chromadb-client==0.5.10
sentence-transformers==2.3.1
openai==1.12.0
python-dotenv==1.0.1
tiktoken==0.6.0
fastapi==0.109.2
uvicorn==0.27.1
